refactor: extract EncryptedCredentialsStore from CloudAuthentication
Pure relocation of the existing EncryptedSharedPreferences/MasterKeys
credential-read logic out of CloudAuthentication into a standalone,
injectable EncryptedCredentialsStore. CloudAuthentication.autoSignIn()
now delegates to it. No behavior change; sets up reuse for the
upcoming auth recovery manager (AC-1043).
